BRANCH INLINING
This makes the history much more linear and puts some of the commits
that cvs2svn sprayed into several branches back as 1 commit onto the
mainline.

This actively diverts some of the commits off the SVN tags and branches
and causes a diff against git. The SVN branches and tags that I'm
"stealing" the commits from are deemed not really worthy by yours truly,
however.

As an exchange, the history of e.g. the ntpd vendor or ee(1) are
extended a bit, so there's that.